Give Button an id named "save"
In the activity_main.xml file, the ViewGroup is composed with 2 View components (EditText and Button). The Button component now has an id ("save") associated with it. The Button's id is stored in an auto-generated Java class named "R" (it stores a bunch of final variables). The Button's id will be used to reference the button object within MainActivity.java.