feat: Write EditText's content to internal storage
Previously, the "save" button was merely displaying predefined text to Logcat as a verification that the OnClickListener was working. Now the "save" button gets a reference to the EditText (View component) instance and writes its content to a file named "note_squirrel.txt" (which is saved to the device's internal storage).