feat: Add load functionality
During app start up (the onCreate() method call), the content of a file (saved/written during a prior "save" button click) is read and subsequently appended to the EditText (a View component) object.

FUTURE-EXERCISE: Try creating a  new Button component, name it "load" and attach an id to it, lay it out to the right of the "save" button, set an OnClickListener, and move the loadSavedFile() call out of the onCreate() and into the implementation of OnClickListener's onClick() method.