refactor: Replace hard-coded String instances with keys from String Resources
Change text that are to be used visually (is rendered to the screen) from hard-coded String objects (which are immutable in Java) to use key-value pairing from String Resources feature (found in "/res/values/strings.xml".

NOTE: This is better practice than simply hard-coded String because it'll be like changing one reference's value instead of having to change each hard-coded usage if we have that same sequence-of-characters in multiple places throughout the project. The String Resources feature also allows for something called Localization (which I'm interpretting as something used for translating our app to different languages). The "strings.xml" file can be editted manually (the raw xml file), but Android Studio also has a "Translations Editor" take can be used to edit the key-value pairing.