feat: Utilize persistent data to determine if saved file exist
To decide if loadSavedFile() should be called at app start-up, there is now a check for whether or not the "save" button was pressed during a previous "run" of the app.

This REQUIRES the app storing data in between "runs" (and we achieve this using "getPreferences()", which save and retrieve persistent key-value pairs of primitive data types) (this is a DIFFERENT data storage option from writing a file to the device's internal storage [e.g. how the EditText's content is stored when the "save" button is clicked]).

If the "save" button was clicked during a previous "run", we use the ACTIVITY's "preference file" to "put()" a key-value(primitive_type) pair AND "commit()" it. Then this persistent data (stored in the activity's "preference file") will be checked the next time the app is started up (when "onCreate()" is called).