feat: Add new Activity (will become image viewer)
ImageActivity is currently an empty Activity, but soon the ViewGroup for its xml file (instruction for laying out components) will have a new View component of type ImageView.

This will become a kind of lock-screen where user will need to touch the screen at particular screen-coordinates in a particular sequence. And if they successfully "unlock" the screen, this Activity will transition to the MainActivity screen (the one with the EditText and "save" Button).