feat: Add ImageView View to starting Activity
The starting Activity (ImageActivity class, with its associated activity_image.xml layout file) is now using a ViewGroup of type LinearLayout. This ViewGroup had a new View component (an ImageView) added to it. The ImageView was given an id named "touch_image" and was specified an image to display (the image named "image_default.png" can be found in the "res/mipmap-xhdpi" directory) using the attribute "android:src".

The app's MANIFEST file was changed to have ImageActivity's (the starting Activity) screenOrientation fixed to "portrait". Also, a previous comment/note was (somehow) lost, and is now reinserted.