feat: Add Dialog message to app's starting Activity
Needed a builder, set options (OK/Cancel buttons, title, and the actual message/instruction prompt to the user) on the builder, have the builder "create" the Dialog instance, and "show" the Dialog instance to the device's screen.

Could have added additional functionality to the user clicking the "OK" (and/or "Cancel") button, but intentionally left it empty/blank because we don't need additional functionality in this case).