feat: Store x and y values to a table in SQLite database
The SQL command to create the table and its specified fields (columns) was refactored into using String.format() (pairing a format-String with format-specifiers [the column names, which are now private constants]). The private constants (column names) are used as keys inside of Database.storePoints(List<Point>) method, where we used a ContentValues object as a data-carrier for the SQL command to INSERT new entries into the database's table.