feat: Retrieve values from a table in SQLite database
Add new method with signature/identifier "getPoints()" to Database class. A SQL command is used to read entries from a table of our database.

We select all entries from columns "x" and "y" from the table identified as "points" from the SQLite database, and have it ordered by column "id" (the column that's specified as PRIMARY KEY). This returns a Cursor object from the package "android.database", which we use to iterate through the values returned by the SQL query. The values are added to an ArrayList<Point> (we verify that this method is working in ImageActivity.pointsCollected() [which gets called by PointCollector using an "Observer" design pattern]).