refactor: Use different thread when writing to database
AsyncTask class is being used to run the writing-to-database code on a separate thread. A dialog alerts the user that writing is in progress (we use Thread.sleep() to slow down that separate thread so we can actually see the dialog before it gets dismissed [our particular task is actually really fast, but using a separate thread will be a good idea for tasks that will take a long time to finish]).