feat: Implement part of ImageActivity.verifyPasspoints(List<Point>)
Display a dialog to user to alert them that the 4 points they touched are being compared to the passcode stored in the SQLite database. The actual comparison will take place in a separate thread of execution (and the actual comparison logic will be implemented in the next commit), which has an overridden "onPostExecute(Boolean)" method that will clear the list of points the user just touched, dismiss the dialog ("Checking passpoints..."), and BASED ON WHAT IS RETURNED FROM "doInBackground(Void...)" (where the actual comparison logic will go)... THE "onPostExecute(Boolean)" WILL EITHER TRANSITION TO MainActivity OR DISPLAY A Toast SAYING "Access Denied" to the screen.

NEXT: The actual comparison of points entered to the points stored in the SQLite database from a previous run.