Implement verifyPasspoints(List<Point>) logic
After prompting user to set a passpoint sequence and storing that sequence in the SQLite database, the user can then attempt to "access" the app by entering the same passpoint sequence (had an allowable distance of error set to 40, but mom was bad at remembering WHERE [on the image] she touched, so I had to increase the allowable-"off-ness" to 100).

If the user is able to touch 4 points (in the correct sequnce) that is "close-enough" to the stored passpoint sequence, then the app will allow the user access to the note-taking screen (the starting activity [ImageActivity] will change to MainActivity).