feat: Implement RESETTING PASSPOINTS menu item
ImageActivity is no longer storing persistent data (has the passpoints been set or not) restricted to its own context (e.g. the stored data is no longer accessible solely in ImageActivity... it use to be MODE_PRIVATE... now its scope is accessible from anywhere within the NoteSquirrel application [such as from MainActivity, where the menu item to RESET THE PASSPOINTS is accessed]). This was achieved by replacing references to "getPreferences(MODE_PRIVATE)" with "PreferenceManager.getDefaultSharedPreferences(getApplicationContext())". Also, ImageActivity had lines of code to create a dialog prompt about storing passpoints moved from its "onCreate()" to "onStart()" (which is not called when an ImageActivity is instantiated, but is called whenever the ImageActivity instance becomes visible... so both when it is first created AND when it is brought back to the screen [e.g. after MainActivity calls "finish()"]).