feat: Change menu item from "Options" to "Take a Photo"
Currently, requesting (via an Intent) to have the package manager query the device for an app that can handle a request to take photo. If there exist an app that can take photo, a subactivity is launched (the photo capturing app).

NEXT: Saving and displaying the photo captured, which requires writing to the device's external storage directory.