refactor: Display image selected from "Browse the Gallery"
PathViewerActivity is launched by MainActivity's menu item "Browse the Gallery". It displays the photo selected (the data returned from the sub-activity) and the Uri path of the photo selected.

The purpose of this (and the previous) commit was to get a better understanding of the differences between Uri path and (absolute) file path... Not understood yet.