refactor: Restructure Game class to use State-design-pattern
Also, COLLISION-DETECTION now checks the left-TOP and left-BOTTOM of ball instead of left-CENTER (the same for COLLISON-DETECTION of the right side of ball; checks for top and bottom corners instead of center).

Sprites (bats and ball) are now smaller in terms of screen size. The ball's x and y movement speeds are increased. The (opponent) bat's simulated intelligence has a higher chance of making smart-choices (moving towards the ball).