Add TilesetUtilActivity (tinker with custom ImageView)
SoundPool only plays the first 5 seconds of a sound file. Switch to MediaPlayer, which will play the sound file in its entirety (then switched from a large [300+ MB] sound file, which caused a long install time, to a smaller sound file).

Had trouble getting TilesetUtilActivity to be inflated by my custom ImageView (TilesetPaletteView) because I was initially using the wrong overloaded constructor, which took time to troubleshoot, but now seems to be loading as expected.