feat: Hook tiles (solid and nonsolid) with movement
BUGGY! The point of this was to try to get more clue on the background movement bug. The control condition for movement is not working properly, but is at a point where each diagonal corner of the screen controls separate direction (left, right, up, down), unlike before where one touch event caused vertical AND horizontal movement at the same time.

BUG: solid and walkable tiles are not exactly where they're suppose to be... which may not actually be the issue... this may be the previous GameCamera bug (potentially related to pixelToScreenRatio).