fix: Discover root of background vs player movement bug
Player.render() had its screen rectangle dependent on the player's xCurrent and yCurrent... its screen placement was changing each time the player's position changed... this screen rectangle SHOULD BE fixed at center of the screen while the xCurrent and yCurrent should be used for BACK-END to determine whether the player's movement will bump into solid tiles or not.

Had to change GameCamera and Scene lines of code to be FIXED IN PLACE in order to narrow down the cause of this bug. HAVE NOT REVERTED BACK TO WHAT THEY REALLY SHOULD BE YET. JUST HAPPY THE MOVEMENT OF BACKGROUND AND FOREGROUND (player's sprite being centered on the screen) IS WORKING PROPERLY! ALSO, SOLID vs NONSOLID ALSO SEEMS TO BE WORKING PROPERLY.