fix: Movement bug for right and bottom edge of sprite fixed
When specifying coordinate points to check whether a tile is solid (e.g. when the player sprite is about to move), an adjustment has to be made for the RIGHT and BOTTOM edge of the player sprite (subtract 1 pixel from the actual coordinate value... I think this may be related to the LEFT and TOP starting at 0).

The player use to get "stuck" when its RIGHT or BOTTOM edge were touching a solid tile... by -1px when checking RIGHT and BOTTOM, the player no longer exhibit this getting "stuck" glitch.