fix: On-screen touch cmoving player properly
Instead of dividing the game screen into 2 pieces horizontally and 2 pieces vertically, it's now being divided into 3 pieces horizontally and 3 pieces vertically. Adjusted the conditions used to translate the MotionEvent's x and y position into left, up, right, and down. The previous attempt (the 2-piece version) wasn't able to separate out horizontal and vertical movement properly; it was enough to move the player sprite around the scene, but it was buggy/not-as-intended (there were 4 quadrants... the top-left moved the player left, the top-right moved the player up, the bottom-left moved the player down, and the bottom-right moved the player right... or something similar to that). Now, on-screen touch events respond more like a directional pad.