feat: Add "swap game" button to switch GameCartridge
In GameView.switchGame(boolean), the presently running thread is shutdown, a new instance of GameCartridge is instantiated, and a new thread will run.

The previous "run" of PoohFarmerCartridge is not saved and therefore the user is starting over from the beginning of the game. This is also true for PongCartridge (but isn't as significant because "runs" of this game is much shorter).

The screen's user input layout doesn't change when in PongCartridge (there's still a directional pad and the "swap game" button in the bottom half of the screen (this game is no longer full screen).