refactor: Set-up to learn Activity/Fragment lifecycle
Override methods related to lifecycle state and add Logcat output for each method. Comment-out TileMap initialization verifier (Logcat output).

The goal is to save the player's position during onPause() and NOT have the game restart to its starting state if the SurfaceView loses focus (its Task is moved to the Back Stack or the screen orientation/layout is changed) and is resumed later on (regain focus).