refactor: Move instantiation of GameCartridge to JackInActivity
GameView.surfaceCreated(SurfaceHolder) will call JackInActivity's gameCartridge's init(SurfaceHolder, int) and loadSavedState() prior to instantiating a GameRunner to run the game cartridge on a separate thread.

Also, the inputs from ButtonPadFragment and DirectionalPadFragment are no longer going through GameView, but are going directly to JackInActivity's gameCartridge.

NEXT: remove the viewport's launching the DVD library activity when touching the bottom-left and bottom-right of the screen, also rework the game switching feature.