refactor: Rework DirectionalPadFragment determining touch event out of bound unsuccessful
Still buggy... holding down a press and then moving it off the button (out of the ImageView's bound) is still NOT stopping the player from moving in that direction.

Direct viewport presses (touching the SurfaceView screen) is implemented to do one move per press (justPressed). I wanted the DirectionalPadFragment to function differently. I want the directional pad to be able to continously move the player while the button (ImageView) is being held down, and to stop when the user take their finger off or drag their finger out of the button's bounds. Currently, unable to achieve this. Currently, the player moves while the press is held down and dragged around (even when the drag/touch is out of bound of the button), and only stops when the user take their finger off.