fix: DirectionalPadFragment no longer getting stuck in one direction
DirectionalPadFragment is performing as intended. Finger does not need to be lifted from screen for changes in direction. Direction selected no longer triggering the previous-direction-button-selected (I think this was related to InputManager's boolean for directional-pad buttons never getting switched-back-to false in between new touch events) (another possibility, though less likely, is the touch event may be out of bound for all 4 direction buttons BUT is still triggering isPressed to be true because it's on the background of DirectionalPadFragment [red background... up-left, up-right, center, down-left, down-right] WHICH MEANS AN ACTUAL "direction" is never set [it stays "null"] but is still performing a callback to InputManager with "isPressed" as "true").

BUG FIXED, DirectionalPadFragment's front-end is performing as intended... back-end NEEDS a cleaning. Also, I think I messed up the viewport's input when the screen orientation is switched to horizontal.