feat: Port tile collision detection for PocketCritters
Successfully brought in and translated tile collision detection feature from December 2019's PC version of PocketCritters. The code was translated to use Android's version of images (instead of BufferedImage, it's Bitmap... same for cropping subimages and finding rgb values). Also, instead of create a tile map from the full world map (due to memory constraints), this version successfully uses a cropped version of the full map (which will have to somehow transition to another cropped version later on [allow the player to walk from one small map to another small map]). The load time is pretty bad and there's still a huge fear of memory issues, but I didn't have to hand-create an rgb-blue-print and was able to re-use the effort-put-in from December's project!