refactor: Generate TileType[][] tiles using external file instead of parsing image
MUCH FASTER LOADING! Instead of parsing the world map image for each run, currently using something similar to rgbTileFarm (an external source file that defines which tile is SOLID or WALKABLE; but instead of pixel, this is using String).

The external file is created from IntelliJ's version of PocketCritters-SerialCritterNabbing because Andio Studio wouldn't let me write out to my netbook's storage.

The external file is the FULL WORLD MAP, so I needed to use "Arrays.copyOfRange()" to crop only what is needed (because I'm using a cropped image of the full world map).