tinker: Test a method tthat returns cropped map IMAGE
The tile-collision TEXT-FILE was divided into 10 sections by 10 sections (each section is 28 tiles by 27 tiles) in the previous commit. This commit is testing a method that takes in int values (between 1-10) for column and row of the desired section (for the starting town, this would be column=3, row=8) and returns a cropped-IMAGE of that section.

This commit is just testing that the method is functioning properly within a sandbox activity (SpriteSheetViewerActivity). Have NOT incorporated into the PocketCrittersCartridge yet. Currently, PocketCrittersCartridge is using a hard-coded (arbitrary) TileMap rather than one of the 10-by-10 section (for both its front-end IMAGE and its back-end tile-collision TEXT-FILE).