tinker: Enable browsing cropped IMAGE of world map sections
The sandbox activity (SpriteSheetViewerActivity) now has directional button (and a center [home] button) to change which section (the full map was divided into 10 columns and 10 rows) is displayed in the ImageView. There is also a TextView to display the current column and row represented by the cropped image.

This was to help verify that the cropping of the sections was done correctly, but it was also suppose to help think about/figure out how to connect the section (how to implement the transition between sections... as we'll as the issue of being in a corner of a section so TWO-neighboring-sections may have to be displayed for seamlessness [may give up on trying for seamlessness and just do a simple transition at first... player walks onto an EDGE-tile and a transition effect will shift the background and place the player on the corresponding EDGE-tile of the new section]).