refactor: Use variables instead of hard-int-values to initialize TileMap.tiles
Decide to NOT use 10 by 10 sections approach to break down FULL world map IMAGE and TileType[][] (tile collision detection map).

Will still break down the FULL world map IMAGE file, but will use the (previously referred to) arbitrary-values approach (to now be referred to as the "natural" approach).

This natural approach avoids section-transition and the problem of multiple-neighboring-sections-needing-to-be-displayed-concurrently-when-in-the-corner-of-a-section.

Reverting back to this approach really mean breaking down the FULL world map IMAGE based on storyline parts (https://bulbapedia.bulbagarden.net/wiki/Appendix:Yellow_walkthrough).