refactor: Separate viewport width and height
Still using a square viewport, but all horizontal (x and width related) and vertical (y and height related) usages have been separate (previous version used one value named sideSquareScreen). Now the values used are named widthViewport and heightViewport (which is effectively still sideSquareScreen); these are different from "widthScreen" and "heightScreen" (which is not a square, but the actually device's screen dimension).

The purpose of this separate is to allow FUTURE IMPLEMENTATION of switching to FULL SCREEN (rather than square viewport) for certain scenes involving the AIs (MS Clippit).