feat: Introduce SceneManager but need to reorganize
The background IMAGE (texture) and TileMap (tile collision detection NESTED-ARRAY) are now using TILE INDEXES from the same source, but that source is currently being initialized in "TileMap.initSpawnPosition()" and should to be placed else where.

This commit is suppose to help with maintainability as the project acquire more scenes.