refactor: Change PoohFarmer background IMAGE to "on-the-fly" implementation
The farm (spring) texture is no longer pre-loaded into Assets when PoohFarmerCartridge is initialize. Instead, the IMAGE is loaded and cropped when needed (with the sprite sheet being set to null to be garbage-collected, leaving just the cropped version).

Brought in new image assets for PocketCrittersCartridge to use for future implementation of "transfer points" and "BattleState", these images were altered to make pixel coordinates easier to work with if translated into tile indexes.