asset: Generate tile collision source text files
Wrote new method in IntelliJ's PocketCritters project (which tweaked code from December2019) to create new text files representing the tile collision maps for Android Studio's PocketCritters project (for the unimplemented scenes: HOME_01, HOME_02, HOME_RIVAL, and LAB).

These new text files are NOT exactly the same as the one for the world map's tile collision; the new text files have an extra 2 tiles for width and height because the IMAGEs associated with them are UN-BORDERED. Including solid tiles for top, left, right, and bottom prevents the player from walking off the edge of the scene.