fix: Remove border from "tile_home01.txt" and adjust clip dimension
GameCamera now uses CLIP_WIDTH_IN_TILE and CLIP_HEIGHT_IN_TILE of 8 and 8 (rather than 9 and 9) since most of the indoor scenes are 8 tile by 8 tile (this avoid the problem of white/blank space showing and the game camera jumping when the player crosses the middle of the scene (both vertically and horizontally).

Also, it's verified that the border tiles to prevent the player from moving off the edge were redundant... "TileMap.isSolid(int, int)" accounts for moving off the edge of the scene.

NEXT: Remove border for HOME_02, HOME_RIVAL, and LAB. Also, fix spawning position (when the player exits an indoor scene... the player should be positioned where they were standing immediately before entering the indoor scene rather than at the spawning position for the start of the game. Should also clean-up back end for scene transition (it's functioning pretty much as expected on the front end, but feels strongly coupled on the back end.