asset: Crop and verify ROBOT sprites
Refactored TilesetUtilActivity to be a sprite sheet viewer (SpriteSheetViewer2Activity) with 4 buttons that iterate over the column and row indexes of the collection of Bitmap from the cropped sprite sheet.

Practice ConstraintLayout, practice new naming convention of pre-pending member variables with "m", practice more judicious usage of "this", and becoming more comfortable with nested-array indexes in for-loop constructs.

GOAL: teach the player programming concepts (front-load via the ComputerActivity) and have the player practice those concepts (via the GameConsoleActivity) by supplying logic to instances of Robot, and have the Robot defend the player (while the player is farming) from waves of enemies (either other robots or pests/insects trying to destroy the player's crop) (tower-defense-like gameplay... while tilling the soil-tiles, watering the crop-entities, and harvesting the cannabis-items from the crop-entities).