Add action items (listview and gridview) to the app bar
Created my own icon images and attached them to menu items (with attribute "showAsAction") in ListFragmentDvdParent (the dvd library with IMAGE resource [the learning-grid-view version of the dvd library is currently just a bunch of String for titles without images]).

The action items representing "ListView" and "GridView" are presently outputting a toast-message, rather than switching layout.