refactor: Add "favorite" toggling image to ModelDvdFragment's layout
Decided I will re-use ModelDvdFragment's xml layout file for the GridView version of the dvd library. Utilizing "scaleType" attribute, using a FrameLayout as a container, and adding a STAR-icon to the bottom-right corner. At this point, I found STAR-icon image resources and have the off-version set as default.

Currently, there isn't an OnClickListener defined to switch the STAR-icon to an on-version. Also, have not actually re-use this xml layout file to inflate GridView's cells yet.