tinker: Add functionality to "GridView" action item (app bar menu-item icon)
Currently, mixing the data source (e.g. ListView is using the DvdList class [which has image resources], while GridView is using the 525 entries populated from a CSV text file). The plan is to inflate GridView using ModelDvdFragment's xml layout file.

Found a way to get the current index of the ListView so pressing the "GridView" action item can scroll to the most recent position of the "ListView" rather than starting from the beginning of the data source.