refactor: Change package structure for dvd library project
Now there is one unifying "dvdlibrary" package that contains 2 directories ("official" and "roughdraftwithimages"). The "official" one will use the 525 entries from the CSV text file as a data source. The "roughdraftwithimages" will use the DvdList class (an ArrayList<Dvd> that's been populated with 16 place-holder elements) as a data source.

Both "official" and "roughdraftwithimages" will have the feature of switching their view from list to grid and vice-versa.

Currently, the "roughdraftwithimages"'s gridview is being populated with Strings from the CVS text file. Will switch this to use DvdList class as its data source in order to populate each gridview's cell with an image.