fix: SpriteSheetVerifier1Activity no longer changes ImageView's dimension at run time
Removed button's onClick()'s lines of code that adjust the layout_width and layout_height of the ImageView. The advantage: the Activity's measuring and laying-out of its Views can rely on the attributes specified by the xml layout file (primarily "adjustViewBounds" and "scaleType") rather than overriding those with specified values.