feat: Render cursor in State.START_MENU (initial stage)
Rendering 2 cursor images (with vertical difference of 16*scaleFactor [scaleFactor in this case is 6]). The second cursor image is to figure out how much to increment the yOffset by when changing the menu item index (which has not yet been included).

NEXT: move the cursor image according to menu item index.