feat: Implement user input for PocketCrittersCartridge's StartMenuState's cursor
The cursor moves when the user presses the up or down buttons on either the viewport or the directional-pad.

This leads to saving/loading due to user specified menu item input VERSUS saving/loading due to device's configuration change (screen rotation, keyboard entry, action that causes the Activity to restart).