feat: Handle dvd library (rough-draft) orientation change
Saving and restoring data (the dvd's "isFavorite" property, which determines which image to use for the "star" [blue or grey]) when the Activity is destroyed and re-created during orientation/configuration change.

Needed to save and restore the Activity's mode (Mode.LIST or Mode.GRID) because the Activity is initialized as a list by default. If the mode was Mode.GRID prior to orientation change... restoring needs to switch to the grid layout.