feat: Add context menu for ViewportFragment's GameView
This is an alternative way to access start-menu menu items (the first way was pressing the menu button from the ButtonPadFragment [where the a-b buttons are also located]). Now, long-pressing the viewport, in areas that do not trigger movement, will pull up the start menu.

This was tricky because I didn't handle "onTouch(View, MotionEvent)" properly in InputManager. I needed to return false when the touch event doesn't get handled by my overwritten method (which I didn't do... I was returning true everytime). Now, I'm only returning true for touch events inside the bounds that triggers movement, and return false if we reach the end of the overwritten method. Returning false tells the system that I didn't handle the touch event, so the event should be propagated up the hierarchy so something further up (e.g. JackInActivity) can receive the touch event. In this case, JackInActivity will receive the touch event, recognize that it's a long-press, and create a context menu for the GameView.

https://developer.android.com/guide/topics/ui/menus#context-menu

https://stackoverflow.com/questions/6179060/how-to-get-ontouchevent-long-click-and-context-menu-working-together