tinker: Replace conventional context menu with dialog
According to "https://stackoverflow.com/questions/38872585/how-to-specify-the-background-color-of-the-context-menu", I can't customize the context menu (I wanted a 2-line border with rounded corners), but can use a dialog for a context menu.

I created a drawable resource file and defined a shape, then applied it as the background for the ViewGroup of the dialog's inflated layout. (followed "https://stackoverflow.com/questions/30219120/view-with-a-border").