fix: Height of each element no longer filling parent's height
RecyclerView is using a GridLayoutManager to determine the data to populate each ViewHolder with (as well as their position), but was laying them out in an unintended manner (each ViewHolder height filled the parent). It was a silly mistake/oversight, but has been corrected (from "match_parent" to "wrap_content") and is now looking much more grid-like.

SIDE-NOTE: Launching and returning from (with proper loading) a new Activity is a BIG DEAL... I'm no longer limited to working within my GameView (an extension of SurfaceView)... I can connect ComputerActivity, TelevisionActivity, start menu, inventory/backpack, exercise/assignments (lessons to be taught from ComputerActivity), etc and return to the present state of the game! This opens up a lot of the built-in features of the Android OS instead of having to rig things up from scratch. Dragging-and-dropping, animated transitions... a bunch of things I haven't learned yet!