feat: Add save/load functionality for PocketCrittersCartridge's alternative start menu
Previously, a contextual menu was added (accessed via long-pressing the viewport). Now, the "load" and "save" menu items of the context menu actually loads and saves the state of the game (it does the same thing as the primary start menu's "load" and "save" menu items that can be accessed via the button-pad's "menu" button).

NEXT: Have the television and computer tiles launch (and return from) their respective Activity.