feat: Pass "Player.inventory" into BackpackActivity AND get clicked "position" as result
Currently, only viewport's context menu is able to pass in the player's inventory (ArrayList<Item>)... StartMenuState's access to BackpackActivity is still using the arbitrary list of items that was used for testing.

BackpackActivity passes back the "position" of the item clicked (except for the case when the user clicked the OS back-button). Have NOT used the passed back value yet, only verifying that it's actually returning a "position" with a debug-Log-message.