fix: Both start menus pop/dismiss after menu-item selected
Previously, selecting menu-item "Save" or "Load" would result in the start menu still being present on the screen (it was hard to tell that it actually saved). This was true for the primary start menu (StartMenuState) and the alternative start menu (viewport's context menu [custom dialog]).

Now, all menu-items will clear (pop/dismiss) the start menu from the screen.

BUG: something is wrong with saving/loading related to StateManager (I think I may have GameState being pushed onto StateManager.stateStack multiple time instead of just being at the base of the stack).