refactor: Incorporate ScenePong into PongCartridge
PongCartridge is now composed with a StateManager. The StateManager has a GameState that is composed with a SceneManager. This SceneManager is using ScenePong.

This is just the half-way point of the goal. Still have to move lines of code out of PongCartridge and into ScenePong (e.g. for "update()" and "render()"). Also, have to convert from using Sprite subclasses to Entity subclasses. May have to subclass Player.