fix: Scene.enter() changes game camera's widthClipInPixel and heightClipInPixel
Each concrete subclass of Scene can specify the clip size (in number of TILES) that it needs in its constructor, which is used (along with tile map's tileWidth and tileHeight) in "Scene.enter()" to set the game camera's widthClipInPixel and heightClipInPixel.

Also, all Entity instances in the scene's entityManager gets their "init()" called when Scene.enter() is called.