feat: Add saving/loading of Entity
Previously, Scene's entity manager was "transient". Now it's not.

Changes to how Entity initializes its "Rect bounds" was needed. Also, needed to change where to instantiate an entity (now in a scene's constructor instead of its "init()" [when the game is loaded from a saved file, we don't want to inflate the collection of entities AND add a newly instantiated entity; we only want to inflate the saved collection of entities]).

EntityManager had an anonymous class that sorts the order of the entities arraylist, which couldn't be serialized. It was changed to an inner-class that implements Comparator and Serializable.