feat: Change HUD's item image through selection from BackpackActivity
Currently, all scenes (and all games) are displaying the HUD. This shouldn't be the case for Pong and Frogger (also, Frogger is displaying the HUD's item image improperly/smaller-than-it-should-be).

Saving/loading when returning from another Activity caused a lot of difficulties, but I think most of them have been fixed.