feat: Add rendering ability to tile framework
Through PocketCritter's game console tile, PoohFarmer's SceneHothouse can respond to a-button press when the player is facing a PotTile. This toggles PotTile.isEmptyTable and sets Tile.image depending on "isEmptyTable".

The result is seeing an image of a pot on the table or just the empty table (in this case, just setting Tile.image to null, which means "render()" won't draw anything to the screen at that tile position).