refactor: Tilling requires shovel to be player's selected item
Also, ordering matters for a-button presses in SceneFarm (e.g. entity-check before tile-check, and if entity-check has a collision then we escape without checking tiles... this effectively makes it one a-button press will produce one action; the other situation is activating the Robot entity and simultaneously tilling the GrowableGroundTile with one a-button press).