feat: New item (SEED_BAG) can instantiate CropEntity
If the GrowableGroundTile isTilled and its cropEntity is null, the seed bag will call the GrowableGroundTile's "plantCropEntity(CropEntity.Id)" method; which sets its cropEntity and adds that cropEntity to the current scene's entityManager.

CropEntity.render(Canvas) is implemented and entity-collision detection is working properly between the player and the newly instantiated seedling.

Moved Item-logic out of Scene.getInputButtonPad() and encapsulated it into Item.execute(Tile).