refactor: Seed is part of tile-system instead of entity-system
Previously, entity-collision (even with bounds set to zeroes) was active for top-left corner of the CropEntity.

Now that seed is a part of the tile-system, tile-collision check is for Tile.Walkability walkability rather than entity-collision check (which is based off the entity's xCurrent and yCurrent; this lead to the top-left corner bug).

NEXT: need to detect "sleeping" through the bed in SceneHouseLevel01 (and that the tile had been watered) to trigger transitioning from seed-to-sproutling (the tile will instantiate an instance of CropEntity, set that instance to its cropEntity field, and add that instance to the scene's entityManager).