feat: First Tile-to-CropEntity instantiation by going to bed
After 2 failed sessions, finally found why this wasn't working. BedTile.execute(GameCartridge) is suppose to check each tile from SceneFarm's tilemap. Instead of querying the CURRENT "xIndex" and "yIndex", I was using "column" and "row" (which are the MAXIMUM "xIndex" and "yIndex" of that tilemap... and returns a "null" to prevent checking out-of-bounds tiles) as the argument for TileMap.getTile(int, int).

Now, CropEntity is being instantiated where GrowableGroundTile was SEEDED and isWatered was true (entity-collision detection working as well)!