refactor: Change EmptyHandsItem to GlovedHandsItem
GlovedHandsItem.execute() is instantiating a Product. Player is composed with a Holdable field that update(long) and render(Canvas) when Holdable is not null.

FIXED: ShippingBinTile's loading bug (gameCartridge was null after loading because super wasn't being called in ShippingBinTile.init(GameCartridge)).

NEXT: check CropEntity to see if it can "regrow" or else set its active to false AND set the GrowableGroundTile's reference to CropEntity as null, implement Player's dropping-a-Holdable (onto an empty tile, as well as into a shipping bin tile).