refactor: Add dropping Product into ShippingBinTile
Successfully placing the harvested Product into the ShippingBinTile.

Currently, if the Product is dropped onto a GrowableGroundTile that doesn't have a CropEntity, the player's Holdable is set to null and the Product "disappear" (is no longer being rendered by the player) and has no collision detection (so the player can't pick the Product back up).

NEXT: compose Scene with an ItemManager, convert Product into a subclass of Item, establish Item collision detection, implementing picking up and Item that's been dropped onto a GrowableGroundTile.

NEXT: still need to handle CropEntity that are regrowable when it gets harvested by the GlovedHandsItem.