feat: Control background tint fx using in-game time
Scene uses a background image with a yellowish-tint when the in-game time reaches 3pm. It uses a background image with a bluish-tint when the in-game time reaches 6pm.

FIX: Instead of tinting (through LightingColorFilter) during each iteration of Scene.render(Canvas) (there was a noticeable slow-down to the game), the Scene class now stores background images when Scene.enter() is called (one for each of TimeManager.ModeOfDay [DAYLIGHT, TWILIGHT, NIGHT]). Now the tinting operation should only happen just one time (during Scene.enter()) instead of the many times Scene.render(Canvas) is called.